Added  Multi-User Registration and Password Hashing
Problem: Only one user could be registered because we were storing only
a single variable in the session.
Solution: Decided to use a Hashmap to register multiple-users. Wrote
LoginUtil and HashUtil classes to assist in this. The LoginUtil class
follows the singleton pattern exposing a getInstance() method.

Problem: Unsafe to store passwords.
Solution: Decided to hash the user's password together with a salt and
to validate a password we compare hashed input against a previously
recorded hash.

Problem: The scope of a hashtable/hashmap is the life of the program
itself. This causes the registered users to be lost when the application
is terminated.
Solution: We serialize objects and store them by writing to a file.
Keeping a file on disk is not very secure but for the scope of this
program we can consider this as a replacement for db.