Fixed redirection
Upon not logged in state if we tried to access a members-only page. The
system was giving an exception. This was because the
session.getAttribute("is_login") returned null when session variable was
not set and we were applying .equals() method on that inside If
condition. Now we have used OR statements to check first using
'short-circuiting' if the session.getAttribute is not returning null
only then it goes to checking the next condition.